feat(mood): move structs to modules and test triangle shape
