+++ crates/clogbox-clap/src/lib.rs
@@ -22,11 +22,11 @@ use clack_extensions::audio_ports::PluginAudioPorts;
 use clack_extensions::note_ports::PluginNotePorts;
 use clack_extensions::params::PluginParams;
 use clack_extensions::state::PluginState;
+pub use clack_plugin;
 pub use clack_plugin::clack_export_entry;
 pub use clack_plugin::entry::SinglePluginEntry;
 use clack_plugin::host::HostMainThreadHandle;
 pub use clack_plugin::host::HostSharedHandle;
-pub use clack_plugin;
 pub use clack_plugin::plugin;
 pub use clack_plugin::plugin::features;
 use clack_plugin::prelude::*;
