+++ crates/clogbox-module/src/context.rs
@@ -263,7 +263,9 @@ impl<E: Enum, T> ops::IndexMut<E> for AudioStorage<E, T> {
 
 impl<T> AudioStorage<Empty, T> {
     /// Empty buffer containing no elements.
-    pub const EMPTY: Self = Self { storage: EnumMapArray::CONST_DEFAULT };
+    pub const EMPTY: Self = Self {
+        storage: EnumMapArray::CONST_DEFAULT,
+    };
 }
 
 /// Storage for events associated with enum channels.
