+++ examples/mood/Cargo.toml
@@ -27,3 +27,6 @@ mood.workspace = true
 az.workspace = true
 log.workspace = true
 num-traits.workspace = true
+
+[dev-dependencies]
+insta = { workspace = true, features = ["csv"] }
